+++ VerticalWarehouses/Panel PC UI/Ferretto.VW.App.Modules/Installation/ViewModels/Elevator/HorizontalResolutionCalibrationViewModel.cs
@@ -871,7 +871,7 @@ namespace Ferretto.VW.App.Installation.ViewModels
 
                     if (!isOk)
                     {
-                        this.ShowNotification(string.Format(Localized.Get("ServiceMachine.InconsistencyStateAndSensorsBay"), otherBay.Number), NotificationSeverity.Warning);
+                        this.ShowNotification(string.Format(Localized.Get("ServiceMachine.InconsistencyStateAndSensorsBay"), (int)otherBay.Number), NotificationSeverity.Warning);
                     }
                 }
             }
